requirements.txt
langchain-openai
langchain_community
langchain-huggingface
python-dotenv
pypdf
langchain-text-splitters
sentence-transformers
langchain-chroma
gradio